pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otpilot"
version = "2.0.0"
description = "Background desktop utility that copies OTPs from Gmail to clipboard on hotkey trigger."
readme = "docs/README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "OTPilot Contributors"},
]
keywords = ["otp", "gmail", "clipboard", "hotkey", "automation", "2fa"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Utilities",
]
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth>=2.23.0",
    "pynput>=1.7.6",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "pyperclip>=1.8.2",
    "plyer>=2.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.urls]
Homepage = "https://github.com/otpilot/otpilot"
Documentation = "https://github.com/otpilot/otpilot/tree/main/docs"
Repository = "https://github.com/otpilot/otpilot"
Issues = "https://github.com/otpilot/otpilot/issues"

[project.scripts]
otpilot = "otpilot.main:main"

[tool.setuptools.packages.find]
include = ["otpilot*"]
